scripts/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
boto3>=1.34.0
PyYAML>=6.0
